Answers refactored, added favorites, little fix for users auth

+++ src/com/itmencompany/servlets/services/DeleteAnswerService.java
@@ -7,11 +7,10 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-
 import com.itmencompany.datastore.dao.IncomingInfoDao;
-import com.itmencompany.datastore.dao.UserOrderDao;
+import com.itmencompany.datastore.entities.AppUser;
 import com.itmencompany.datastore.entities.IncomingInfo;
-import com.itmencompany.datastore.entities.UserOrder;
+import com.itmencompany.helpers.AppUserHelper;
 
 @WebServlet("/delete_answer")
 public class DeleteAnswerService extends HttpServlet {
@@ -29,15 +28,22 @@ public class DeleteAnswerService extends HttpServlet {
 
 	protected void doPost(HttpServletRequest request, HttpServletResponse response)
 			throws ServletException, IOException {
+		String res = "ok";
+		AppUser appUser = AppUserHelper.getUserFromRequest(request);
+		if(appUser == null){
+			response.sendRedirect("/login");
+			return;
+		}
+	
 		String answer_id = request.getParameter(ANSWER_ID_KEY);
 
 		response.setCharacterEncoding("utf-8");
 		response.setContentType("text");
 		PrintWriter pw = response.getWriter();
 		IncomingInfoDao dao = new IncomingInfoDao();
-		String res = "ok";
+		
 		try {
-			if (answer_id != null) {
+			if (answer_id != null && appUser.getIsAdmin()){
 				IncomingInfo info = dao.get(Long.parseLong(answer_id));
 				dao.delete(info);
 			}
